refactor(adapter-oas): carve resolvers.ts and factory.ts into model/, load/, emit/
resolvers.ts mixed four unrelated concerns in 604 lines. Split by phase:

- model/components.ts: getSchemas, sortSchemas, collectRefs, extractSchemaFromContent
- model/operations.ts: getParameters, getRequestSchema, getResponseSchema, the
  content-type helpers, and their private response-ref-resolving helpers
- model/server.ts: resolveBaseUrl, resolveServerUrl
- emit/schemaHelpers.ts: flattenSchema, getDateType, getSchemaType,
  getPrimitiveType, isHandledFormat, extractExamples
- emit/createNode.ts: gains buildSchemaNode, its only remaining caller

factory.ts's bundle/upgrade/validate pipeline moves to load/, split into
load/source.ts (read, fetch, bundle) and load/normalize.ts (parseDocument,
parseFromConfig, validateDocument), with load/index.ts as the import
surface. operation.ts is untouched; model/operations.ts sits next to it.

Tests split to match: resolvers.test.ts into model/components.test.ts,
model/server.test.ts, emit/schemaHelpers.test.ts, and emit/createNode.test.ts;
factory.test.ts moves to load/normalize.test.ts (every function it covers
now lives there).

Part of the adapter-oas leaner-shape refactor (slice 06/07): no behavior
change, verified by the characterization snapshots.
