+++ packages/adapter-oas/src/adapter.ts
@@ -4,11 +4,12 @@ import type { AdapterSource } from '@kubb/core'
 import { DEFAULT_PARSER_OPTIONS } from './constants.ts'
 import { buildDiscriminatorChildMap, patchDiscriminatorNode } from './emit/discriminator/index.ts'
 import type { DiscriminatorTarget } from './emit/discriminator/index.ts'
-import { assertInputExists, parseDocument, parseFromConfig, validateDocument } from './factory.ts'
+import { assertInputExists, parseDocument, parseFromConfig, validateDocument } from './load/index.ts'
+import { getSchemas } from './model/components.ts'
+import { resolveBaseUrl } from './model/server.ts'
 import { getOperations } from './operation.ts'
 import { createSchemaParser } from './parser.ts'
 import { collectInlineEnums, refPromotedEnums } from './promoteEnums.ts'
-import { getSchemas, resolveBaseUrl } from './resolvers.ts'
 import { reportSchemaDiagnostics } from './schemaDiagnostics.ts'
 import type { AdapterOas, Document, SchemaObject } from './types.ts'
