+++ packages/adapter-oas/src/emit/createNode.ts
@@ -1,6 +1,7 @@
 import { ast } from '@kubb/ast'
-import { buildSchemaNode } from '../resolvers.ts'
+import { extractExamples } from './schemaHelpers.ts'
 import type { SchemaContext } from './parseSchema.ts'
+import type { SchemaObject } from '../types.ts'
 
 /**
  * The `schema`/`name`/`nullable`/`defaultValue` slice of a context, the only part
@@ -13,6 +14,24 @@ type NodeBaseContext = Pick<SchemaContext, 'schema' | 'name' | 'nullable' | 'def
  * {@link createNode} stays in sync with the AST layer without redeclaring the union.
  */
 type CreateSchemaProps = Parameters<typeof ast.factory.createSchema>[0]
+
+/**
+ * Collects the shared metadata fields passed to every `createSchema` call.
+ */
+export function buildSchemaNode(schema: SchemaObject, name: string | null | undefined, nullable: true | undefined, defaultValue: unknown) {
+  return {
+    name,
+    nullable,
+    title: schema.title,
+    description: schema.description,
+    deprecated: schema.deprecated,
+    readOnly: schema.readOnly,
+    writeOnly: schema.writeOnly,
+    default: defaultValue,
+    examples: extractExamples(schema),
+    format: schema.format,
+  } as const
+}
 
 /**
  * Builds a schema node from a converter's base context plus its type-specific fields. Every
