+++ packages/adapter-oas/src/emit/parseSchema.ts
@@ -1,11 +1,11 @@
 import type { ast } from '@kubb/ast'
 import { isBinary, isReference } from '../oas.ts'
 import type { Refs } from '../refs.ts'
-import { isHandledFormat } from '../resolvers.ts'
 import type { Document, SchemaObject } from '../types.ts'
 import { convertAllOf, convertMultiType, convertRef, convertUnion } from './converters/composition.ts'
 import { convertBinary, convertBoolean, convertConst, convertEnum, convertFormat, convertNumeric, convertString, createNullNode } from './converters/scalar.ts'
 import { convertArray, convertObject, convertTuple } from './converters/structural.ts'
+import { isHandledFormat } from './schemaHelpers.ts'
 
 /**
  * Pre-computed per-schema context passed to every schema converter.
