--- packages/adapter-oas/src/resolvers.ts
@@ -1,605 +1,0 @@
-import { pascalCase } from '@internals/utils'
-import type { ast } from '@kubb/ast'
-import { Diagnostics } from '@kubb/core'
-import { formatMap, SCHEMA_REF_PREFIX, specialCasedFormats, structuralKeys } from './constants.ts'
-import { isJsonMimeType, isReference } from './oas.ts'
-import { getRequestContent, getResponseByStatusCode } from './operation.ts'
-import { createRefs, dereferenceWithRef, derefInPlace } from './refs.ts'
-import type { ContentType, Document, MediaTypeObject, Operation, ParameterObject, ResponseObject, SchemaObject, ServerObject, ServerOptions } from './types.ts'
-
-/**
- * Reads the server URL from the document's `servers` array at `server.index`,
- * interpolating any `server.variables` into the URL template.
- *
- * Returns `null` when `server.index` is omitted or out of range.
- *
- * @example Resolve the first server
- * `resolveBaseUrl({ document, server: { index: 0 } })`
- *
- * @example Override a path variable
- * `resolveBaseUrl({ document, server: { index: 0, variables: { version: 'v2' } } })`
- */
-export function resolveBaseUrl({ document, server }: { document: Document; server?: ServerOptions }): string | null {
-  const index = server?.index
-  const entry = index !== undefined ? document.servers?.at(index) : undefined
-
-  return entry?.url ? resolveServerUrl(entry, server?.variables) : null
-}
-
-/**
- * Replaces `{variable}` placeholders in an OpenAPI server URL with provided values.
- * Resolution order: `overrides[key]` → `variable.default` → left unreplaced.
- * Throws if an override value is not in the variable's `enum` list.
- *
- * @example
- * ```ts
- * resolveServerUrl(
- *   { url: 'https://{env}.api.example.com', variables: { env: { default: 'dev', enum: ['dev', 'prod'] } } },
- *   { env: 'prod' },
- * )
- * // 'https://prod.api.example.com'
- * ```
- */
-export function resolveServerUrl(server: ServerObject, overrides?: Record<string, string>): string {
-  if (!server.variables) {
-    return server.url
-  }
-
-  let url = server.url
-  for (const [key, variable] of Object.entries(server.variables)) {
-    const value = overrides?.[key] ?? (variable.default != null ? String(variable.default) : undefined)
-    if (value === undefined) {
-      continue
-    }
-
-    if (variable.enum?.length && !variable.enum.some((e) => String(e) === value)) {
-      throw new Diagnostics.Error({
-        code: Diagnostics.code.invalidServerVariable,
-        severity: 'error',
-        message: `Invalid server variable value '${value}' for '${key}' when resolving ${server.url}. Valid values are: ${variable.enum.join(', ')}.`,
-        help: `Use one of the allowed enum values, or drop the enum on the '${key}' server variable.`,
-        location: { kind: 'document', pointer: '#/servers' },
-      })
-    }
-
-    url = url.replaceAll(`{${key}}`, value)
-  }
-
-  return url
-}
-
-/**
- * Returns the Kubb `SchemaType` for a given OAS `format` string, or `null` if not found.
- * Formats not in `formatMap` (e.g., `int64`, `date-time`) are handled separately by parser options.
- */
-export function getSchemaType(format: string): ast.SchemaType | null {
-  return formatMap[format as keyof typeof formatMap] ?? null
-}
-
-/**
- * Whether the parser maps `format` to a dedicated type. True for any `formatMap` entry, plus the
- * `specialCasedFormats` that `convertFormat` handles directly. False means the format falls back to
- * the base type, which is what `KUBB_UNSUPPORTED_FORMAT` flags. Reading both sources keeps the
- * diagnostic in step with the parser as `formatMap` grows.
- */
-export function isHandledFormat(format: string): boolean {
-  return getSchemaType(format) !== null || specialCasedFormats.has(format)
-}
-
-/**
- * Converts an OAS primitive type string to its `PrimitiveSchemaType` equivalent.
- * Numeric types (`number`, `integer`, `bigint`) pass through unchanged. `boolean` maps to `'boolean'`. Everything else becomes `'string'`.
- */
-export function getPrimitiveType(type: string | undefined): ast.PrimitiveSchemaType {
-  if (type === 'number' || type === 'integer' || type === 'bigint') return type
-  if (type === 'boolean') return 'boolean'
-
-  return 'string'
-}
-
-export type OperationsOptions = {
-  contentType?: ContentType
-}
-
-/**
- * Returns all parameters for an operation, merging path-level and operation-level entries.
- * Operation-level parameters override path-level ones with the same `in:name` key.
- * Each `$ref` parameter is dereferenced via `dereferenceWithRef` before merging.
- *
- * @example
- * ```ts
- * getParameters(document, operation)
- * // [{ name: 'petId', in: 'path', required: true, schema: { type: 'integer' } }]
- * ```
- */
-export function getParameters(document: Document, operation: Operation): Array<ParameterObject> {
-  const resolveParams = (params: Array<unknown>): Array<ParameterObject> =>
-    params.map((p) => dereferenceWithRef(document, p)).filter((p): p is ParameterObject => !!p && typeof p === 'object' && 'in' in p && 'name' in p)
-
-  const operationParams = resolveParams(operation.schema?.parameters || [])
-  const pathItem = document.paths?.[operation.path]
-  const pathLevelParams = resolveParams(pathItem && !isReference(pathItem) && pathItem.parameters ? pathItem.parameters : [])
-
-  const paramMap = new Map<string, ParameterObject>()
-  for (const p of pathLevelParams) {
-    if (p.name && p.in) {
-      paramMap.set(`${p.in}:${p.name}`, p)
-    }
-  }
-  for (const p of operationParams) {
-    if (p.name && p.in) {
-      paramMap.set(`${p.in}:${p.name}`, p)
-    }
-  }
-
-  return Array.from(paramMap.values())
-}
-
-function getResponseBody(responseBody: boolean | ResponseObject, contentType?: string): MediaTypeObject | false {
-  if (!responseBody) return false
-  if (isReference(responseBody)) return false
-
-  const body = responseBody as ResponseObject
-  if (!body.content) return false
-
-  if (contentType) {
-    if (!(contentType in body.content)) return false
-    return body.content[contentType]!
-  }
-
-  const contentTypes = Object.keys(body.content)
-  const availableContentType = contentTypes.find(isJsonMimeType) ?? contentTypes[0]
-  if (!availableContentType) return false
-
-  return body.content[availableContentType]!
-}
-
-// Dereference every response `$ref` in place so callers can read response bodies directly.
-function resolveResponseRefs(document: Document, operation: Operation): void {
-  const responses = operation.schema.responses as Record<string, unknown> | undefined
-  if (!responses) return
-  for (const key in responses) {
-    derefInPlace({ document, container: responses, key })
-  }
-}
-
-/**
- * Returns the response schema for a given operation and HTTP status code.
- *
- * Returns an empty object `{}` when no response body schema is available.
- *
- * @example
- * ```ts
- * getResponseSchema(document, operation, 200)         // SchemaObject
- * getResponseSchema(document, operation, '4XX')       // {}
- * ```
- */
-export function getResponseSchema(document: Document, operation: Operation, statusCode: string | number, options: OperationsOptions = {}): SchemaObject {
-  resolveResponseRefs(document, operation)
-
-  const responseBody = getResponseBody(getResponseByStatusCode({ document, operation, statusCode }), options.contentType)
-
-  if (responseBody === false) {
-    return {}
-  }
-
-  const schema = responseBody.schema
-
-  if (!schema) {
-    return {}
-  }
-
-  return dereferenceWithRef(document, schema)
-}
-
-/**
- * Returns the request body schema for an operation, or `null` when absent.
- *
- * @example
- * ```ts
- * getRequestSchema(document, operation) // SchemaObject | null
- * ```
- */
-export function getRequestSchema(document: Document, operation: Operation, options: OperationsOptions = {}): SchemaObject | null {
-  if (operation.schema.requestBody) {
-    operation.schema.requestBody = dereferenceWithRef(document, operation.schema.requestBody)
-  }
-
-  const requestBody = getRequestContent({ document, operation, mediaType: options.contentType })
-
-  if (requestBody === false) {
-    return null
-  }
-
-  const mediaType = Array.isArray(requestBody) ? requestBody[0] : options.contentType
-  const schema = Array.isArray(requestBody) ? requestBody[1].schema : requestBody.schema
-
-  // OAS 3.1 (and the 3.0 -> 3.1 upgrade) drops the schema for an `application/octet-stream` body,
-  // leaving an empty media type object. Synthesize the binary schema so generators still emit a
-  // request body type for the operation.
-  if (mediaType === 'application/octet-stream' && (!schema || Object.keys(schema).length === 0)) {
-    return { type: 'string', contentMediaType: 'application/octet-stream' }
-  }
-
-  if (!schema) {
-    return null
-  }
-
-  return dereferenceWithRef(document, schema)
-}
-
-/**
- * The three component sections Kubb reads schemas from.
- */
-type SchemaSourceMode = 'schemas' | 'responses' | 'requestBodies'
-
-/**
- * A schema annotated with its component section source and original name. `getSchemas` uses this
- * to resolve name collisions across sources.
- */
-type SchemaWithMetadata = {
-  schema: SchemaObject
-  source: SchemaSourceMode
-  originalName: string
-}
-
-export type GetSchemasOptions = {
-  contentType?: ContentType
-}
-
-export type GetSchemasResult = {
-  schemas: Record<string, SchemaObject>
-  /**
-   * Maps a renamed component pointer (`#/components/<source>/<name>`) to the
-   * collision-resolved unique name used as the key in `schemas`. Components that keep
-   * their original name are not recorded, so the map stays empty for documents
-   * without collisions.
-   */
-  renames: Map<string, string>
-}
-
-/**
- * Returns `true` when `fragment` carries any JSON Schema keyword that makes it
- * structurally significant on its own (see `structuralKeys`).
- *
- * A fragment with a structural keyword can't be safely merged into a parent schema.
- */
-function hasStructuralKeywords(fragment: SchemaObject): boolean {
-  for (const key in fragment) {
-    if (structuralKeys.has(key as 'properties')) return true
-  }
-  return false
-}
-
-/**
- * Flattens a keyword-only `allOf` into its parent schema.
- *
- * Only flattens when every member is a plain fragment, with no `$ref` and no structural keywords
- * (see `structuralKeys`). Outer schema values take precedence over fragment values.
- * Returns `null` for a `null` input, and the original schema unchanged when flattening is unsafe.
- *
- * @example
- * ```ts
- * flattenSchema({ allOf: [{ description: 'A pet' }], type: 'object', properties: {} })
- * // { type: 'object', properties: {}, description: 'A pet' }
- * ```
- *
- * @example
- * ```ts
- * flattenSchema({ allOf: [{ $ref: '#/components/schemas/Pet' }] })
- * // returned unchanged, contains a $ref
- * ```
- */
-export function flattenSchema(schema: SchemaObject | null): SchemaObject | null {
-  if (!schema?.allOf || schema.allOf.length === 0) return schema ?? null
-
-  const allOfFragments = schema.allOf as Array<SchemaObject>
-  if (allOfFragments.some((item) => isReference(item))) return schema
-  if (allOfFragments.some(hasStructuralKeywords)) return schema
-
-  // Destructure `allOf` out instead of `delete merged.allOf`: a `delete` transitions the freshly
-  // spread object into V8 dictionary (slow) mode, and this runs per `allOf` schema during parsing.
-  const { allOf: _allOf, ...rest } = schema
-  const merged = rest as SchemaObject
-
-  for (const fragment of allOfFragments) {
-    for (const [key, value] of Object.entries(fragment)) {
-      merged[key as keyof SchemaObject] ??= value as SchemaObject[keyof SchemaObject]
-    }
-  }
-
-  return merged
-}
-
-/**
- * Extracts the inline schema from a media-type `content` map.
- *
- * Prefers `preferredContentType` when given, otherwise uses the first key in the map.
- * Returns `null` when `content` is absent, the schema is missing, or the schema is a `$ref`.
- *
- * @example
- * ```ts
- * extractSchemaFromContent(operation.content, 'application/json')
- * // SchemaObject | null
- * ```
- */
-export function extractSchemaFromContent(content: Record<string, unknown> | undefined, preferredContentType?: ContentType): SchemaObject | null {
-  if (!content) return null
-
-  const firstContentType = Object.keys(content)[0] ?? 'application/json'
-  const targetContentType = preferredContentType ?? firstContentType
-  const contentSchema = content[targetContentType] as { schema?: SchemaObject } | undefined
-  const schema = contentSchema?.schema
-
-  if (schema && '$ref' in schema) return null
-  return schema ?? null
-}
-
-/**
- * Walks a schema tree and collects the names of all `#/components/schemas/<name>` `$ref`s.
- */
-function* collectRefs(schema: unknown): Generator<string, void, undefined> {
-  if (Array.isArray(schema)) {
-    for (const item of schema) yield* collectRefs(item)
-    return
-  }
-
-  if (schema && typeof schema === 'object') {
-    for (const key in schema) {
-      const value = (schema as Record<string, unknown>)[key]
-      if (!(key === '$ref' && typeof value === 'string')) {
-        yield* collectRefs(value)
-        continue
-      }
-      if (value.startsWith(SCHEMA_REF_PREFIX)) {
-        const name = value.slice(SCHEMA_REF_PREFIX.length)
-        if (name) yield name
-      }
-    }
-  }
-}
-
-/**
- * Returns a copy of `schemas` topologically sorted by `$ref` dependency.
- *
- * Referenced schemas appear before the schemas that depend on them, so code generators
- * can emit types in the correct order. Cycles are silently skipped.
- *
- * @example
- * ```ts
- * const sorted = sortSchemas({ Order: orderSchema, Pet: petSchema })
- * // Pet appears before Order when Order.$ref points at Pet
- * ```
- */
-export function sortSchemas(schemas: Record<string, SchemaObject>): Record<string, SchemaObject> {
-  const deps = new Map<string, Array<string>>()
-
-  for (const [name, schema] of Object.entries(schemas)) {
-    deps.set(name, [...new Set(collectRefs(schema))])
-  }
-
-  const sorted: Array<string> = []
-  const visited = new Set<string>()
-
-  function visit(name: string, stack: Set<string>) {
-    if (visited.has(name) || stack.has(name)) return
-    stack.add(name)
-    for (const child of deps.get(name) ?? []) {
-      if (deps.has(child)) visit(child, stack)
-    }
-    stack.delete(name)
-    visited.add(name)
-    sorted.push(name)
-  }
-
-  for (const name of Object.keys(schemas)) {
-    visit(name, new Set())
-  }
-
-  const result: Record<string, SchemaObject> = {}
-  for (const name of sorted) result[name] = schemas[name]!
-  return result
-}
-
-const semanticSuffixes: Record<SchemaSourceMode, string> = {
-  schemas: 'Schema',
-  responses: 'Response',
-  requestBodies: 'Request',
-}
-
-/**
- * Collects component schemas from one or more sources and resolves name collisions.
- *
- * Sources default to `['schemas', 'requestBodies', 'responses']`. Returned schemas are
- * topologically sorted by `$ref` dependency so generators emit types in the correct order.
- *
- * When two or more schemas normalize to the same PascalCase name:
- * - Same source → numeric suffix (`2`, `3`, …).
- * - Different sources → semantic suffix (`Schema`, `Response`, `Request`).
- *
- * @example
- * ```ts
- * const { schemas, renames } = getSchemas(document, { contentType: 'application/json' })
- * ```
- */
-export function getSchemas(document: Document, { contentType }: GetSchemasOptions): GetSchemasResult {
-  const components = document.components
-  const refs = createRefs(document)
-
-  function resolveSchemaRef(schema: SchemaObject): SchemaObject {
-    if (!isReference(schema)) return schema
-    const resolved = refs.resolve<SchemaObject>(schema.$ref)
-    return resolved && !isReference(resolved) ? resolved : schema
-  }
-
-  const candidates: Array<SchemaWithMetadata> = [
-    ...Object.entries((components?.schemas as Record<string, SchemaObject>) ?? {}).map(([name, schema]) => ({
-      schema: resolveSchemaRef(schema),
-      source: 'schemas' as const,
-      originalName: name,
-    })),
-    ...(['responses', 'requestBodies'] as const).flatMap((source) =>
-      Object.entries(components?.[source] ?? {}).flatMap(([name, item]) => {
-        const schema = extractSchemaFromContent((item as { content?: Record<string, unknown> }).content, contentType)
-        return schema
-          ? [
-              {
-                schema: resolveSchemaRef(schema),
-                source,
-                originalName: name,
-              },
-            ]
-          : []
-      }),
-    ),
-  ]
-
-  const normalizedNames = new Map<string, Array<SchemaWithMetadata>>()
-  for (const item of candidates) {
-    const key = pascalCase(item.originalName)
-    const bucket = normalizedNames.get(key) ?? []
-    bucket.push(item)
-    normalizedNames.set(key, bucket)
-  }
-
-  const schemas: Record<string, SchemaObject> = {}
-  const renames = new Map<string, string>()
-
-  for (const [, items] of normalizedNames) {
-    const isSingle = items.length === 1
-    let hasMultipleSources = false
-    if (!isSingle) {
-      const firstSource = items[0]!.source
-      for (const item of items) {
-        if (item.source !== firstSource) {
-          hasMultipleSources = true
-          break
-        }
-      }
-    }
-
-    items.forEach((item, index) => {
-      const suffix = isSingle ? '' : hasMultipleSources ? semanticSuffixes[item.source] : index === 0 ? '' : String(index + 1)
-      const uniqueName = item.originalName + suffix
-      schemas[uniqueName] = item.schema
-      if (suffix) renames.set(`#/components/${item.source}/${item.originalName}`, uniqueName)
-    })
-  }
-
-  return { schemas: sortSchemas(schemas), renames }
-}
-
-/**
- * Resolves the AST type descriptor for a date/time format, honoring the `dateType` option.
- * Returns `null` when `dateType: false`, so the format falls through to `string`.
- */
-export function getDateType(
-  options: ast.ParserOptions,
-  format: 'date-time' | 'date' | 'time',
-): { type: 'datetime'; offset?: boolean; local?: boolean } | { type: 'date' | 'time'; representation: 'date' | 'string' } | null {
-  if (!options.dateType) {
-    return null
-  }
-
-  if (format === 'date-time') {
-    if (options.dateType === 'date') {
-      return { type: 'date', representation: 'date' }
-    }
-    if (options.dateType === 'stringOffset') {
-      return { type: 'datetime', offset: true }
-    }
-    if (options.dateType === 'stringLocal') {
-      return { type: 'datetime', local: true }
-    }
-    return { type: 'datetime', offset: false }
-  }
-
-  if (format === 'date') {
-    return {
-      type: 'date',
-      representation: options.dateType === 'date' ? 'date' : 'string',
-    }
-  }
-
-  // time
-  return {
-    type: 'time',
-    representation: options.dateType === 'date' ? 'date' : 'string',
-  }
-}
-
-/**
- * Collects the shared metadata fields passed to every `createSchema` call.
- */
-/**
- * Reads schema examples as an array. OAS 3.1 uses an `examples` array, but specs (including ones
- * labeled 3.1) still use the singular OAS 3.0 `example`, which the upgrader only converts on the
- * 3.0 -> 3.1 hop. Normalize both into one array so the AST node exposes only `examples`.
- */
-export function extractExamples(schema: SchemaObject): Array<unknown> | undefined {
-  if (Array.isArray(schema.examples)) return schema.examples
-  return schema.example !== undefined ? [schema.example] : undefined
-}
-
-export function buildSchemaNode(schema: SchemaObject, name: string | null | undefined, nullable: true | undefined, defaultValue: unknown) {
-  return {
-    name,
-    nullable,
-    title: schema.title,
-    description: schema.description,
-    deprecated: schema.deprecated,
-    readOnly: schema.readOnly,
-    writeOnly: schema.writeOnly,
-    default: defaultValue,
-    examples: extractExamples(schema),
-    format: schema.format,
-  } as const
-}
-
-/**
- * Returns all request body content type keys for an operation.
- *
- * The requestBody is dereferenced in place when it is a `$ref` (the same mutation that
- * `getRequestSchema` already performs), so the returned list accurately reflects the
- * available content types even for referenced bodies.
- *
- * @example
- * ```ts
- * getRequestBodyContentTypes(document, operation)
- * // ['application/json', 'multipart/form-data']
- * ```
- */
-export function getRequestBodyContentTypes(document: Document, operation: Operation): Array<string> {
-  if (operation.schema.requestBody) {
-    operation.schema.requestBody = dereferenceWithRef(document, operation.schema.requestBody)
-  }
-
-  const body = operation.schema.requestBody as { content?: Record<string, unknown> } | undefined
-  if (!body) return []
-
-  // dereferenceWithRef keeps $ref but spreads all resolved fields (including `content`).
-  // Do not bail out on isReference, the content is already present on the merged object.
-  return body.content ? Object.keys(body.content) : []
-}
-
-/**
- * Returns all response content type keys for an operation at a given status code.
- *
- * Response `$ref`s are resolved in place first (the same mutation `getResponseSchema` performs),
- * so the returned list reflects the available content types even for referenced responses.
- *
- * @example
- * ```ts
- * getResponseBodyContentTypes(document, operation, 200)
- * // ['application/json', 'application/xml']
- * ```
- */
-export function getResponseBodyContentTypes(document: Document, operation: Operation, statusCode: string | number): Array<string> {
-  resolveResponseRefs(document, operation)
-
-  const responseObj = getResponseByStatusCode({ document, operation, statusCode })
-  if (!responseObj || typeof responseObj !== 'object' || isReference(responseObj)) return []
-
-  const body = responseObj as { content?: Record<string, unknown> }
-  return body.content ? Object.keys(body.content) : []
-}
